+++ src/ui/rect.rs
@@ -1,14 +1,14 @@
 #[derive(PartialEq, Eq, Hash, Copy, Clone, Debug)]
 pub struct Rect {
-    pub x                   : u32,
-    pub y                   : u32,
-    pub width               : u32,
-    pub height              : u32,
+    pub x                   : usize,
+    pub y                   : usize,
+    pub width               : usize,
+    pub height              : usize,
 }
 
 impl Rect {
 
-    pub fn new(x: u32, y: u32, width: u32, height: u32) -> Self {
+    pub fn new(x: usize, y: usize, width: usize, height: usize) -> Self {
         Self {
             x               : x,
             y               : y,
@@ -17,7 +17,7 @@ impl Rect {
         }
     }
 
-    pub fn new_with_size(width: u32, height: u32) -> Self {
+    pub fn new_with_size(width: usize, height: usize) -> Self {
         Self {
             x               : 0,
             y               : 0,
@@ -35,7 +35,7 @@ impl Rect {
         }
     }
 
-    pub fn is_inside(&self, p: (u32, u32)) -> bool {
+    pub fn is_inside(&self, p: (usize, usize)) -> bool {
         p.0 >= self.x && p.1 >= self.y && p.0 < self.x + self.width && p.1 < self.y + self.height
     }
